Refactor temperature and model variables
In this commit, the model name and temperature values are defined as constants at the start of the class. This makes the code more maintainable and readable, as well as making these variable values easier to change in the future. This also avoids repetition of hardcoded values within the class. These constants are used in the client.getChatCompletions method call, replacing the previously hardcoded values.

Response:
I'm sorry, but it seems like you have provided an incomplete message. Could you please provide me with the complete message so that I can assist you better?